style: apply every remaining ruff autofix (safe and unsafe)
2076 -> ~830 findings. Everything ruff can rewrite mechanically, plus the
executable-bit and shebang findings:

- EXE001/EXE002: 34 files got the executable bit their shebang asks for, 15
  lost a bit they had no shebang for.
- EXE005: pa_mqa_logits.py's 66 `#!=----` separators are decorative comments,
  not shebangs -- file-level noqa rather than rewriting them all.
  hipify_python.py had a real shebang below the licence header; moved to line 1.
- RUF013's 38 sites were pre-checked for the trap that bit batch 2 twice: every
  annotation involved is a genuine builtin type (int/float/str), no factory
  functions, so making the Optionals explicit is safe.

One autofix had to be reverted: TC004 hoisted `from fla import __version__`
out of its `if TYPE_CHECKING:` block in gated_delta_rule_utils.py. fla is an
optional dependency, so that turned `import aiter.ops.triton.gated_delta_net`
into a ModuleNotFoundError. The guard is restored with a noqa explaining why.
(The runtime read of __version__ under a TYPE_CHECKING-only import is a
pre-existing latent bug in this vendored file; not touched here.)

Import sweep back at the 249 ok / 1 pre-existing fail baseline, black clean,
all 357 touched files compile.

+++ aiter/ops/triton/comms/iris.py
@@ -106,7 +106,7 @@ def calculate_heap_size(
 
     # Total with overhead
     total_bytes = mem_input + mem_rs + mem_quant + mem_gather
-    total_with_overhead = int(math.ceil(total_bytes * overhead_factor))
+    total_with_overhead = math.ceil(total_bytes * overhead_factor)
 
     logger.debug(
         f"Heap size calculation: M={M}, N={N}, dtype={dtype}, world_size={world_size}, "
